+++ Cargo.toml
@@ -1,33 +1,13 @@
-workspace = { members = ["cli/ga4gh-sdk-cli"] }
-[package]
-name = "ga4gh-sdk"
+[workspace]
+members = [
+    "cli",
+    "lib"
+]
+
+[workspace.package]
 version = "0.1.0"
-authors = ["Aarav Mehta <aarav05mehta@gmail.com>", "Pavel Nikonorov", "ELIXIR Cloud & AAI <cloud-service@elixir-europe.org>"]
+authors = ["Aarav Mehta <aarav05mehta@gmail.com> (Google Summer of Code Participant)", "Pavel Nikonorov <pavel@genxt.network> (Google Summer of Code Mentor; GENXT)", "ELIXIR Cloud & AAI <cloud-service@elixir-europe.org> (ELIXIR Europe)"]
 edition = "2021"
-description = "Generic SDK and CLI for GA4GH API services"
-license = "Apache-2.0"
+readme = "./README.md"
+license-file = "LICENSE.txt"
 repository = "https://github.com/elixir-cloud-aai/ga4gh-sdk.git"
-
-[dependencies]
-tokio = { version = "1", features = ["full"] }
-serde = "^1.0"
-serde_derive = "^1.0"
-serde_json = "^1.0"
-url = "^2.2"
-uuid = { version = "^1.0", features = ["serde", "v4"] }
-log = "0.4"
-env_logger = "0.9"
-once_cell = "1.8.0"
-
-[dependencies.reqwest]
-version = "^0.11"
-features = ["json", "multipart"]
-
-[dev-dependencies]
-mockito = "0.31"
-mockall = "0.10.2"
-cargo-nextest = "0.9.30"  
-
-[lib]
-name = "ga4gh_sdk"
-path = "lib/src/lib.rs"
