+++ cli/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "cli"
+description = """
+A cross-platform CLI tool written in Rust to use GA4GH-sdk
+"""
+repository = "https://github.com/elixir-cloud-aai/ga4gh-sdk/tree/main/cli"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+readme.workspace = true
+license-file.workspace = true
+
+[dependencies]
+ga4gh-lib = { path = "../lib" }
+clap = "3.0"
+clap_complete = "3.0"
+tokio = { version = "1", features = ["full"] }
+
+[[bin]]
+name = "cli"
+path = "src/main.rs"
