+++ lib/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "ga4gh-lib"
+description = """
+Generic SDK and CLI for GA4GH API services
+"""
+repository = "https://github.com/elixir-cloud-aai/ga4gh-sdk/tree/main/lib"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+readme.workspace = true
+license-file.workspace = true
+# rust-version = "1.74"
+
+[dependencies]
+tokio = { version = "1", features = ["full"] }
+serde = "^1.0"
+serde_derive = "^1.0"
+serde_json = "^1.0"
+url = "^2.2"
+uuid = { version = "^1.0", features = ["serde", "v4"] }
+log = "0.4"
+env_logger = "0.9"
+once_cell = "1.8.0"
+
+[dependencies.reqwest]
+version = "^0.11"
+features = ["json", "multipart"]
+
+[dev-dependencies]
+mockito = "0.31"
+mockall = "0.10.2"
+cargo-nextest = "0.9.30"  
+
+[lib]
+name = "ga4gh_sdk"
+path = "lib/src/lib.rs"
